% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count-common-slots.R
\name{count_common_slots}
\alias{count_common_slots}
\title{Count Doodle matches}
\usage{
count_common_slots(n_colleagues, o, p, dep)
}
\arguments{
\item{n_colleagues}{number of colleagues}

\item{o}{number of options available}

\item{p}{number of picked options}

\item{dep}{degree of dependency between reviers}
}
\value{
dataframe
}
\description{
Count Doodle matches
}
\examples{
count_common_slots(n_colleagues, o, p, dep)
}
