% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doodle-samples.R
\name{doodle_samples}
\alias{doodle_samples}
\title{Simulate Doodle Samples}
\usage{
doodle_samples(n_colleagues, o, p, r = 1000, dep = 0)
}
\arguments{
\item{n_colleagues}{number of colleagues addressed}

\item{o}{number of options to choose from}

\item{p}{number of picked options}

\item{r}{number of repetitions of simulations (defaults to 1e3)}

\item{dep}{degree of dependency between colleagues (defaults to 0)}
}
\value{
dataframe with samples
}
\description{
Simulates Doodle samples as a basis for computing the match probability.
}
\examples{
doodle_samples(n_colleagues, o, p)
}
