% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-matches.R
\name{plot_matches}
\alias{plot_matches}
\title{Plot the number of Doodle matches}
\usage{
plot_matches(smples_tbl, n_colleagues, p, o, dep)
}
\arguments{
\item{smples_tbl}{the output of the function `summarize_samples`}

\item{n_colleagues}{number of colleagues addressed}

\item{p}{number of picked options}

\item{o}{number of available options}

\item{dep}{degree of dependency between colleagues}
}
\value{
ggplot
}
\description{
Plot the number of Doodle matches
}
\examples{
plot_matches(smples_tbl, n_colleagues, p, o, dep)
}
