% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob-doodle-match.R
\name{prob_doodle_match}
\alias{prob_doodle_match}
\title{Computes the probability of a doodle match}
\usage{
prob_doodle_match(n_colleagues, o, p, r = 1000, dep, return_prob = TRUE)
}
\arguments{
\item{n_colleagues}{Number of colleagues addressed}

\item{o}{number of options to choose from}

\item{p}{number of picked options per colleagues}

\item{r}{number of repetions (simulates) samples}

\item{dep}{dependencies between reviewers}

\item{return_prob}{return a single number?}
}
\value{
dataframe or numeric
}
\description{
Computes the probability of a doodle match
}
\examples{
prob_doodle_match(3, 10, 3, 0)
}
