% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop-match.R
\name{prop_match}
\alias{prop_match}
\title{Extracts probability of a doodle match from summarized results}
\usage{
prop_match(smples_tbl, n_colleagues)
}
\arguments{
\item{smples_tbl}{summarized samples}

\item{n_colleagues}{number of colleagues}
}
\value{
proability of a match (numeric)
}
\description{
This helper function extracts the probability of a match
- given the parameters n,p,o,r -
from the data frame returned by `summarize_samples(smples)`.
}
\examples{
prop_match(smples_tbl, n_colleagues)
}
