% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-samples.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarizes Monte Carlo samples for finding the doodle match probability}
\usage{
summarize_samples(smples)
}
\arguments{
\item{smples}{Output of `estimate_common_slots`}
}
\value{
dataframe with summarized sample characteristics
}
\description{
Summarizes Monte Carlo samples for finding the doodle match probability
}
\examples{
summarize_samples(smples)
}
