% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-grid.R
\name{vis_grid}
\alias{vis_grid}
\title{Visualize the sensitivity grid of a Doodle match analysis}
\usage{
vis_grid(d_grid)
}
\arguments{
\item{d_grid}{output of `populate_grid`}
}
\value{
ggplot object
}
\description{
Visualize the sensitivity grid of a Doodle match analysis
}
\examples{
vis_grid(d_grid)
}
