% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-grid2.R
\name{vis_grid2}
\alias{vis_grid2}
\title{Visualize the sensitivity grid of a Doodle match analysis}
\usage{
vis_grid2(d_grid, facet_var_rows, facet_var_cols)
}
\arguments{
\item{d_grid}{output of `populate_grid`}
}
\value{
ggplot object
}
\description{
Visualize the sensitivity grid of a Doodle match analysis
}
\examples{
vis_grid(d_grid)
}
