% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-common-slots.R
\name{estimate_common_slots}
\alias{estimate_common_slots}
\title{Monte Carlo analysis of doodle matches}
\usage{
estimate_common_slots(
  slots_chosen_collision,
  n_colleagues,
  o,
  p,
  p_sd,
  dep,
  r = 1000
)
}
\arguments{
\item{slots_chosen_collision}{slots chosen by all colleagues (match)}

\item{n_colleagues}{number of colleagues}

\item{o}{number of options to choose from}

\item{p}{number of picked options}

\item{p_sd}{sd of p}

\item{dep}{degree of dependency beteen colleagues}

\item{r}{number of repeated simulations}
}
\value{
dataframe of samples for Doodle match
}
\description{
Essentially a loop around `count_common_slots`
}
\examples{
estimate_common_slots(slots_chosen_collision, n_colleagues, o, p, p_sd, dep)
}
