% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate-grid.R
\name{populate_grid}
\alias{populate_grid}
\title{Runs a sensitivity analysis for doodle mathces}
\usage{
populate_grid(n_colleagues, o, p, dep, verbose = FALSE)
}
\arguments{
\item{n_colleagues}{number of colleagues}

\item{o}{number of options to choose from}

\item{p}{number of picked options}

\item{dep}{degree of dependency}
}
\value{
dataframe with results of Doodle match sensitivity analysis
}
\description{
Runs a sensitivity analysis for doodle mathces
}
\examples{
populate_grid(n_colleagues, o, p, dep)
}
