% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-samples.R
\name{summarize_samples}
\alias{summarize_samples}
\title{Summarizes Monte Carlo samples for finding the doodle match probability}
\usage{
summarize_samples(smples)
}
\arguments{
\item{smples}{Output of `estimate_common_slots`}
}
\value{
dataframe with summarized sample characteristics
}
\description{
The returned dataframe shows for each simulated samples (denoted by `id`),
whether or not there `is_match`. In more detail, if `Freq >= n_colleagues`,
there `is_match`. Which slots happened to be the match (e.g, all colleagues
chose slot 1) is not so important, but reported by `slots_chosen`.
}
\examples{
summarize_samples(smples)
}
